Register and use the generated service worker
Currently, Chrome still honours the cache headers for the service worker file,
which means that when a new version of the app is updated the browser won't know
the service worker has changed because it gets it from the cache.

The browsers have a restriction to cap the cache max-age to 24 hours, but it is
a bit confusing because the file itself is a mechanism for controlling cache.

However, this behaviour is due to change in browsers in the future.
See here for more detail:
https://stackoverflow.com/questions/41422474/new-version-available-with-service-worker-and-sw-precache

Setting `no-cache` for the service worker allows the app to be updated upon
deploy as expected. If there are bandwidth concerns, this could be set to 1-5
minutes (or similar) to act as a rate-limiter but no one uses this app except me
at this stage :P