Allow state to be persisted to localstorage
This means I can show the previously downloaded data while I wait for new data
to be fetched from Firebase. In the meantime, the user should see a less
intrusive loading indicator.

This approach has the added benefit of still showing data while the app is
offline. It is also relatively easy to show the offline status.