<App> to always render a <CheckinTable>
The check for the valid user will happen at the very start of the render method,
this will be where the homepage component will be rendered.

The check for no data has been moved in to the CheckinTable component.
This begs the question if a custom Placeholder component is still required, but
it does encapsulate the "default" usage and styles of the <Paper> component.