Only install the service worker when logged in
This is a courtesy to the user so we don't store the app on their system unless
they are actually signed in and using it.

In the future, the homepage could have more information about the app so
visitors will know what they are signing up for. If they decide against using
it, nothing is stored. If they decide to use it, it'll be stored for offline
usage.

This has the nice side effect that the notifications, etc. don't need to work on
the homepage, so that one-liner to render the homepage if there is no user still
works.

The app component is getting pretty cluttered and has quite a few
responsibilities. If anything else needs to be added, a refactor should probably
be considered.