Use the new firebase account/app
For such a simple app, I don't think I need to have a staging and production
account. This means I would need my build and CI to have configuration for two
accounts when I'm really the only user for now.

If I feel the need to make a staging account in the future, it's pretty easy to
do and now that I own the domain name, I could even add a CNAME for it.