Clean up and remove the Y axis
Don't need the fake data anymore, using real data instead!

I plan on having more than one Y axis and thought it would be easier to start
with a Y axis and then if I really miss if, I can add it back and try to find a
way to make it look good with 3 of them! (or some other number in the future!)

The padding for the chart seems to be an implementation detail so I don't think
the caller needs to worry about passing it as a prop.