Make pinch to zoom work!
For this to work the scales are stored in the components state, including two
copies of the X scale (the actual scale to modify and an original to not modify
but base changes off).

The <rect> element at the end is to make the whole area register pointer events
so the scroll event handler is fired. If an area only has invisible elements
(fill not set) the events are not normally fired.