Rename Catalogue to ImportExport
The main reason for the catalogue was so people could change the categorization
of items on their list.

Ideally this would be done with drag and drop but without delving into react-dnd
yet, the add dialog can change the designated category of an item.
When a user realises they can't drag and drop, they will probably just create
the item they want to change and when the button changes from "add" to "move"
should indicate they have done what they were after - albeit not as intuitively
as dragging.

Therefore a catalogue UI is probably not needed, however the one use case that
can't be handled so far is to remove an item from the auto complete. This would
be where a catalogue would be handy but having a screen where you can copy and
paste raw data in or out will indirectly provide this feature too.