WIP: Get handleAdd to partially work
This currently just writes a new document or updates an existing document in the
items collection, it still needs to do the catalogue entry.

This was the first of two approaches:
1. Use `where` to find the doc to update (or create a new one)
2. Just `put` over a known id, of which we could need to calculate (slug)

The second approach requires us to come up with a key name for the document
instead of using the autogenerated one but gives us the ability to write our
data straight to a document without needing to query first.