Save items with a slug instead of autogenerated ids
Using auto generated ids mean that if I need to check if an item exists to
update it (or otherwise to create a new document) I need to perform a query
where as if I can predictably create slugs I don't need to, I can just overwrite
whatever is there.

Also, this operation happens in conjunction with other operations and so I'd
like to use a transaction, but it seems only `get`, `set`, `update` and `delete`
work within a transaction and not `where`.

I'm new to Firestore so maybe I'm missing something but this seems like an
easier option.

I was also going to use a slugify npm module, but they add a lot of code and I
don't think I'll need all the features (although I might regret that!).
However in my case, these slugs are not even consumer facing so it should be ok
:fingers-crossed: