Make `handleMark` work with a transaction
Docs: https://firebase.google.com/docs/firestore/manage-data/transactions

According to the docs: "All writes execute at the end of a successful
transaction" which means there is no optimistic update when marking an item.

I think I should remove the transaction for this operation because its not super
important that it gets the value from the database before writing the new one
and the user experience is degraded for the majority of the cases with this
transaction.