Setup the listener for the catalogue first
While this isn't guaranteed, this ordering should prevent errors in the
situation where the items have been synced but the catalogue hasn't.

This is because rendering the items depends on the catalogue.

Even if these do sync in the wrong order, a console error will occur and it will
be fixed automatically once the next update comes through. This commit just
keeps the console quieter.