Install and use react-router-dom
This also deletes the "Open Another List" feature because I originally thought
the URL would just redirect/replace with the list URL you wanted when you browse
to the homepage but it makes sense to use URLs as they were intended.

If you want to use multiple lists, its probably easier to just add multiple
lists URLs to the home screen of your phone.

This might change in the future with a more advanced UI but keeping it simple
for now.