Bring back Open Another List and use localStorage
Because when you add this PWA to the homescreen of an Android device, it will
always default to the homepage instead of the path, we have to persist the path
to localStorage and automatically load that list upon visiting.

The Open Another List feature is the only way other than clearing localStorage
that the user can actually get to another list.