Remove the ImportExport feature
Now that the data is stored in FireStore instead of localStorage, this isn't
really needed. While it may be nice to import and export the data from one list
in FireStore to another there is probably little value in that.

TBH, that feature was really just a quick way for me to modify local storage
from my phone anyway :P