Bugfix and upgrade all the things! (react v16, etc)
NOTE: This was mainly to fix a service worker issue where the browser would try
to register a service worked at `/list/service-worker.js` which wouldn't work.
I upgraded to remove me implementation of service worker from sw-precache and
replaced it with the standard one that comes with `create-react-app`.

However, my implementation registered the service worker from the App component
and used a modified registerServiceWorker script so that I could pass callbacks
when the service worker was installed, etc. and pop up a snackbar notification.
This feature is now not present anymore unless I modify the code again - I may
want to bring this back in the future.