Use latest CRA service worker but turn it off
The service worker is preventing updates from happening which I think is due to
new behaviour from CRA where the service worker is not updated until all tabs
are closed, which I don't think works when you save the web app to the
homescreen of a phone as an "app".

Because of this and the fact that this app is likely going to be used online in
most cases, I've disabled the service worker.
This means:
  - The app won't work offline (not sure about Firebase supported anyway)
  - Sadly, there won't be a prompt to add this to the homescreen, but you
    can still add it manually.