Use a form element to improve mobile UX
This makes the enter button on my android phone turn into a "tab" button instead
of an "enter" button on the first field and means the form can be filled out
without having to tap on anything other than the keyboard.

Also put a max-width the lists so it doesn't look so wide on desktop for ease of
reading.