A basic working planner (add and clear)
The added items to planned are not added to the catalogue but this paves the way for a "recipe" concept.

The bottom nav replaces the URL rather than pushing so navigating back and forth don't constantly add to the history.

The sweep button is also only made available if it can be used so it doesn't confuse people when they first start.