Add plan to list UI (without backend working yet)
This UI has it's own local state to work out which items to add to the list but it doesn't include putting it into the previously entered category.
In previous scenarios, the category information also came from the UI but that was because the user had the option change it.
Adding that ability to this UI would probably make it a bit complicated as opposed to the user going back to the list and rearranging it from there.

That might change in the future as there is still the click and hold interaction free.

A couple of minor tweaks in this commit too:
- The bottom nav has a border-top to make it clearer
- The FAB is disabled unless loading is finished