Refactor the world
The main refactor here is extracting the backend into a context provider and come consumer hooks.
This should make it easier to create a component especially to auto complete the item name so it can be used within add and edit for both the list and the planner.

The other refactor is using the named exports from material-ui, I just find this easier to read and write.