Better touch interactions
Replacing the default tap highlighting in chrome on Android with an active color
that fades in.

This also disabled highlighting text in the dialogues so that tapping and
holding doesn't select the text.
This means copying the text in the dialogue will not be possible but the things
that should be copyable are in text fields which shouldn't be effected.

Closes #18