Remove the landing page
Technically I still need a landing page so the JS can check if there is a
previously used listId in localStorage but now it is very subtle and shouldn't
cause a noticable flash when loading for the first time.

If I cared about SEO, this should probably have some information about the app
and be pre-rendered as to not rely on JS but this isn't important to me right
now... maybe one day.

Closes #12